Split node classes into separate files
